Add parser and parse multi commands per line
Functionally, this commit allows the user to put more than one command
per line by separating them with semi colons. However, most of the
changes in this commit are actually organizational. There is now a
simple parser with some tests as well as a function for variable
expansion.

The on_command function now processes any number of commands in a
string. This helps with multiple commands on a line (e.g. `help; ls`).
One drawback to this approach was that most of the return statements in
on_command had to be switched to continue statements making the code
slightly more confusing. I think this is fine because most of those
areas in the code should be reworked anyway.