Remove empty commands `if`, `fi`, and `else`
These commands did nothing and could never be accessed because of how
on_command is written. The only reason I could think to include them
would be to have them show up in help. However, they did not have any
help strings and I don't think the grammar for the language should be a
built into the command that lists the available builtins.