Cleanup most compiler warnings with error messages
Handle Result types and a few minor changes. Only one compiler warning
should be left, which is the use of std::thread::sleep_ms. I am leaving
this for now so that I don't have to go into redox and make sure the
Duration API exists in our clone of std.