Implement Init and `source` Command
The following changes will allow ion to evaluate an init file in the user's home directory, such as **/home/mmstick/.ionrc**. In addition, I've also added the capability to use the `source` command which will either read from a defined file argument as input or re-read the init file. Hopefully you won't mind,  but I've also cleaned up the last bit of code at the end of the main function so that it's a `while let` loop instead of a `loop` loop.