Feature gate summary the var RECORD_SUMMARY
This is an experimental feature that allows a parsable record of the execution
time of serial commands to be stored to file for display later. They are also
currently mixed in the "up arrow" history, but if that is not desired they can
be filted out based on the comment tag of `#summary#`.

Please seek out the author of this commit for answers if you have questions.