Eliminate Some Memory Copies
This will merge the glob expansion method with the existing expansion method so that arguments will not need to be cloned twice. In addition, this will use a draining iterator in the expansion method and skip the expansion check for the first argument. The build_command method is also improved to use a draining iterator instead of checking each argmument.