Validate Syntax in StatementSplitter
Each statement will now be validated for correct syntax. When an
error is found, a descriptive error will be printed to standard
error and that statement will be skipped.

The two types of errors that are curretly detected are unterminated
subshells and unescaped parenthesis in the wrong place. More types
will be added to this logic in the near future, such as unterminated
braces or invalid characters in brace logic.

This is the first step towards an upcoming brace expansion rewrite
that I am planning to perform.