Rewrite Shell Expansion Module
This work required a lot of mental effort to complete, but the results
are more than worth it. Not only is the new code easier to read, manage,
and extend, but it is faster and provides better syntax parsing than
what we had before.

- Rewrote the 'WordIterator'
  - Reimplemented Brace Words
  - Implemented Nested Branches
  - Reimplemented Variable Words
  - Implemented Process Words
  - Reimplemented Tilde Words
  - Reimplemented Normal Words
  - Reimplemented Whitespace Words
  - Reimplemented Quote Handling
  - No more Result error checking
- Rewrote Brace Expansion
  - Implemented Nested Branches
  - Words are now properly collected before expansion
- Rewrote Variable Expansion
  - The WordIterator does all the work here
  - Module removed as it is no longer required
- Shell expansion now correctly returns a vector of strings
- Implemented an ArgumentSplitter for parsing for expressions
  - Does not yet handle parsing processes correctly
  - Shamelessly copied from my Parallel project
- Performance Improved ++

With this out of the way, I will be implementing more features on top of
this really soon. Arrays will become first class citizens soon, and the
syntax validation layer will see much more work.