Implement Length Methods
Implements the `$len()` and `$len_bytes()` methods for strings, and
the `@len()` method for arrays. The len method for arrays returns
the number of elements in the array. The len method for strings
returns the number of graphemes in the string, whereas the
len_bytes method returns the number of bytes the string occupies.