Monitor Child w/ waitpid & Fork Shell w/ BG Jobs
This change will modify the way that job statuses are monitored.
Rather than using the wait() method on a child, this will use
the lower level waitpid() method from the nix crate, enabling
the ability to detect and record changes to the child process
automatically.

This also implements a change that will fork the shell when a
command is executed that needs to be run in the background.
However, these commands still don't work at the moment due to a
parsing issue.