Handle unusual errors during process spawning
Only print "Command not found: ...", if the error is actually that the command
is not found. Otherwise print a general error description from the system. I
found that while trying to echo to the example scheme (:vec) I was getting
EBADF which caused "command not found: /bin/ion" to be printed.