Enable Function Pipelining
A `RefinedJob` may now be either an `External`, `Builtin`, or `Function`. The output of an
executed pipeline'd function may now be piped into the next command. There are some ways
that this could be improved though, such as enabling the last argument to be an array,
and being able to set the exit status of the function.