Remove time output for incomplete command
This removes time output when a command is incomplete. It also has the
side effect of removing time output for time commands in higher
runlevels like if time is inside a for loop. I am uncertain if that is
a good thing, but I belive this fits closer to how people actually use
the time command while removing unhelpful output.