Change time from builtin to shell keyword. (#505)
* Change time from builtin to shell keyword.

Refactors time from a builtin to a shell keyword, enabling it to handle
all commands including shell builtins. The design has the following
components:

 - A Time variant of Statement a boxed statement member.

 - Support for parsing the time keyword in parser/statement/parse.rs by
   recursing on the rest of the input to the parse function when time is
   encountered.

 - A port of the logic from the old builtin into the execute_toplevel
   impl for Shell, where instead of using the error handling from
   Command, the time statement is executed through a recursive call to
   execute_toplevel.

* Remove time output for incomplete command

This removes time output when a command is incomplete. It also has the
side effect of removing time output for time commands in higher
runlevels like if time is inside a for loop. I am uncertain if that is
a good thing, but I belive this fits closer to how people actually use
the time command while removing unhelpful output.

* Add Time match arm to collect_loops.

* Finish time builtin to keyword refactor

Extensively modify flow to be able to recursively handle statements at
almost every level. This is because the Time variant of Statement itself
owns another Statement.

* Fix parsing of argumentless time keyword

The keyword "time" with no arguments is now parsed into
a Statement::Time containing a Statement::Default