Refactor Pipeline to support multiple redirection
Create a `PipeItem` struct that holds a `Job` and vectors of
`Redirection` and `Input` respectively. This means that the statement:

    `cat <in1 <in2 | tr 'a' 'b' ^> err > out`

Is parsed into a `Pipeline` where `items` contains two jobs. One with
two input redirections and one with two output redirections.

TODO: Fix tests and add tests to cover new parsing cases.