Alter tests for refactor.
Mostly change the parsing tests. This also adds one new test to test
multiple redirection parsing.

The parsing change is somewhat breaking. It needs to be decided if that
is desired. For example, the following used to redirect `file` to the
`cat` command.

    ```
    cat | tr 'a' 'b' < file > output
    ```

Now, file would be redirected to the standard input of `tr`. It is an
open question as to whether or not standard output from `cat` and the
redirected input from `file` should be concatenated in the above case,
and if they are concatenated, then in what order.