Fix order of arguments in call to exec_multi_in
Was accidentally passing the argument for stdin to stdout, causing
contents of a pipe to not be passed to the stdin of the next command in
a pipe. Discovered that the pipe field of RefinedJob::Cat was
superfluous, so it has been removed.

Also reordered some arguments to be more consistent with the rest of the
pipe_exec module.