Create fork abstraction w/ cmd expans & fn prompts
This will eliminate some redundancy caused by manually handling forking
logic in two different areas: command expansions, and function prompts.
This replacement takes a closure which provides access to the inner
child's Shell structure, and automatically handles setting up pipes
and returning the standard output as a String.

As an added bonus, this also eliminates the need to copy the PROMPT
function before executing it.