writeln!(stderr, ...) -> eprintln!
While working in other tasks, I noticed that there's a rather
inconsistent way to handle errors. In some places we use `eprintln!` and
in others `writeln!` to standard error.

It seems like the use of `writeln!` don't have an strong motive, as the
output is always to standard error, the lock is always handled the same
than `eprintln` and there's no obvious (to me) optimization, so I
decided to use the more idiomatic `eprintln!` everywhere.