Fix issues with escaping on double quote context
Escaping symbols on double-quote environment was not working as
intended. The preovious code, when it found a backslash on a quote, it
generated a Word which finished with an escape backslash as text.

With the new code, when a backslash is found a word is emitted up to the
backslash. After this, it checks if the following character should be
escaped: If it shouldn't, we don't advance the cursor and we parse that
backslash as text, on a follwoing word. If it should be escaped, we
advance the buffer, to avoid that the backslash is added to the
following word.