Remove Shell::update_variables
The functionality in this method was to update the PWD
and OLDPWD variables when it was detected that there
was a change to the PWD. Rather than executing this
after every command, this will now only happen during
a SUCCESS exit status within the cd builtin.

Fixes a problem where the init script could change
the PWD with the cd command, but the REPL would not
reflect that change at first.