[WIP] Foreground Job Control Overhaul
This resolves a number of significant issues with the way that foreground jobs were being handled. However, let it be known that there is an issue introduced by this commit which causes the piped Ion scripts to fail, due to some outputs getting mixed up. Therefore, be cautious about using Ion for scripts at the moment. I will look into a fix for the issue tomorrow.