[WIP] Pipeline Execution
Processes within a pipeline should have the same PGID. As this was not
the case, this commit will fix that problem. In addition, processes
within a pipeline should be SIGSTOP'd before they begin execution, and
then SIGCONT'd in the parent when the next job in line is also
SIGSTOP'd. This allows commands earlier in the queue to properly receive
the SIGPIPE signal so that commands like `yes | head` will function
properly.

I've also refactored quite a bit of code within the pipeline execution
module, and have rewritten external command execution to use fork +
execve instead of the Command functionality in the standard library. As
every other piece of functionality in Ion depends upon fork+exec
behavior, it is therefore ideal that external command execution also
works in the same way. This will allow me to do further refactoring in
the near future, and to diagnose some issues that I'm having.

NOTE: Ion remains to be broken with this commit, so I don't recommend to
use Ion until this work is complete.