Use CLAP library for argument parsing
Use the specialized CLAP library for argument parsing rather than a hand rolled one. It streamlines the CLI format while handling most of the edge cases for us as well as adding color and suggestions.
It also allows future improvement at a lower cost and improves the separation between the scripting engine and the frontend used to call it.