Move the scopes to a new member crate
Separate the scope datastructure in a member crate.
Will make possible to release it a separate crate.
Better separation of concerns.
By decoupling the variables from the underlying datastructure, it will
hopefully be easier to change the scopes structure to something more
performant.
By using into, it is now possible to pass a &str directly, rather than
having to convert it beforehand.