Remove the GetVariables Trait
The GetVariables Trait encouraged cloning rather than using references.
Since it is rarely the case that cloning is actually needed, it is
better to replace it entirely.
Furthermore, most of its usage included matching, so it does not
significantly increase general verbosity of the code.
To account for the env vars that must be fetched, a new method, get_str,
was introduced.