feat: Add a fake interactive mode, where stdin is not a terminal
IDEs like IntelliJ run "interactive" scripts without a tty. To make it
possible to use Ion without needing to restart it at eache error, add
the '-f' flag, which will use the stdin like a file, but won't exit on
error. This should NOT be used for being lazy with scripts