feat(readline): Use rustyline rather than liner
Rustyline provides better API and is more actively developed than liner.
It already provides support for hint, which could be used with shellac
to provide as-you-type suggestions. It is also a lot more flexible while
avoiding completely the use of heap allocations.

BREAKING CHANGE: This uses a new readline implementation.