feat: Set $CMD_DURATION variable after each command
The variable is equal to the amount of WHOLE seconds spent executing the
previous command.

This variable is currently not set when running the PROMPT function,
because this would mean that the CMD_DURATION would always refer to
the time spent in the PROMPT function when running in interactive mode.