feat: Make arrays able to nest
Prior to this commit, the following type was considered invalid:
    `[[float]]`

Of course, sometimes it is helpful to have arrays within arrays, and as
such, this patch allows for nested arrays. To do this, the array types
in `Primitive` have all been converted to a `Primitive::Array` type,
that holds a `Box<>` to another `Primitive`. There is - of course - a
slight performance penalty to be expected with moving to using another
`Box` for arrays, however after having run the benchmarks, the
difference appears to be negligible.