fix: in terminator removed unsafe by expect on converting bytes into UTF8 str
Reason: There is no sign that scripts for ion to parse are
properly checked if they contain valid UTF-8.

Documented the fact that the current lexing does only handle UTF-8
content.

Renamed module quotes to terminator which is the visible concept for
outside code

Added unit tests for termination of arrays/subshells